python_project/app/requirements.txt
flask==3.2.0
psycopg2-binary==2.9.9
